Expose response to ``/api/sessions``.
This makes the raw JSON response to the "root path" request immediately
available as an attribute of RESTOAuthClient.  This change also caches
the value so subsequent accesses do not require extra HTTP requests.

We don't know how higher-level objects might use this data, and we
shouldn't care.  Let them (ab)use it appropriately.