Disable Session Keepalives in the Request Library
When this library makes web calls, it re-uses connections to the backend
through the Requests ConnectionPool (part of URLLIB3). In a highly
threaded environment where you re-use these connections ALOT (but, with
delays between reuses), RightScale seems to kill off your connections
from under you. This causes very random and hard to reproduce
'BadStatusLine()' exceptions.

It seems much safer to just disable keepalives so that connections are
terminated as soon as they're done handling a request. The additional
overhead is minimal when creating new connections.