Remove explicit dependency on simplejson.
The only reason we had it in there was to catch a simplejson-specific
exception when decoding the JSON from a response object.  We actually
don't care if it's simplejson or not - if there's an error decoding it,
we just want to roll along.

This allows folks to use stdlib json instead of requiring simplejson.