Fix version-reading in setup.py that prevented pip install.
It was impossible to use pip to install the code because the setup.py
file imported rightscale/__init__.py, which subsequently imported the
entire application and all of its dependencies, causing a circular loop.

Since the VERSION tag isn't used anywhere but the setup.py, I've moved
it into its own text file and updated setup.py to read the version from
there. This cleans up the installation via pip and allows for
python-rightscale to be installed via a requirements.txt file easily.